requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0

# Exchange connectivity
ccxt>=4.0
python-binance>=1.0.19
websockets>=12.0

# LLM
anthropic>=0.40.0

# Technical analysis
pandas>=2.0
numpy>=1.24
ta>=0.11.0
mplfinance>=0.12

# Web dashboard
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1

# Channels
python-telegram-bot>=21.0
aiohttp>=3.9

# Logging and utilities
structlog>=24.0
rich>=13.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
